requirements.txt
asgiref==3.4.1
Django==3.2.6
pytz==2021.1
sqlparse==0.4.1
oslo.messaging>=12.8.0
django-money==2.0.1
openstacksdk==0.58.0
djangorestframework==3.12.4
django-filter==2.4.0
Markdown==3.3.4
gunicorn==20.1.0